Smart-Semantic-Search/requirements.txt
sentence-transformers
psycopg2-binary
pgvector
tqdm
